pyproject.toml
[tool.poetry]
name = "payments"
version = "0.1.0"
description = ""
authors = ["Anna Kartasheva <annakartashevamail@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
loguru = "^0.7.0"
uvicorn = "^0.23.1"
black = "^23.7.0"
isort = "^5.12.0"
faker = "^19.1.0"
pydantic = "^1.10"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.3"}
asyncpg = "^0.28.0"



[tool.aerich]
tortoise_orm = "app.db_config.db_config"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
